Move tests into dedicated file
