+++ misc/discv5/src/behaviour.rs
@@ -1077,145 +1077,3 @@ pub enum Discv5Event {
         closer_peers: Vec<PeerId>,
     },
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use enr::EnrBuilder;
-    use libp2p_core::identity;
-    use std::net::Ipv4Addr;
-
-    #[test]
-    fn test_updating_connection_on_ping() {
-        let keypair = identity::Keypair::generate_secp256k1();
-        let ip: IpAddr = "127.0.0.1".parse().unwrap();
-        let enr = EnrBuilder::new("v4")
-            .ip(ip.clone().into())
-            .udp(10001)
-            .build(&keypair)
-            .unwrap();
-        let ip2: IpAddr = "127.0.0.1".parse().unwrap();
-        let keypair2 = identity::Keypair::generate_secp256k1();
-        let enr2 = EnrBuilder::new("v4")
-            .ip(ip2.clone().into())
-            .udp(10002)
-            .build(&keypair2)
-            .unwrap();
-
-        // Set up discv5 with one disconnected node
-        let mut discv5: Discv5<Box<u64>> =
-            Discv5::new(enr, keypair.clone(), ip.into(), false).unwrap();
-        discv5.add_enr(enr2.clone());
-        discv5.connection_updated(enr2.node_id().clone(), None, NodeStatus::Disconnected);
-
-        let mut buckets = discv5.kbuckets.clone();
-        let mut node = buckets.iter().next().unwrap();
-        assert_eq!(node.status, NodeStatus::Disconnected);
-
-        // Add a fake request
-        let ping_response = rpc::Response::Ping {
-            enr_seq: 2,
-            ip: ip2,
-            port: 10002,
-        };
-        let ping_request = rpc::Request::Ping { enr_seq: 2 };
-        let req = RpcRequest(2, enr2.node_id().clone());
-        discv5
-            .active_rpc_requests
-            .insert(req, (Some(1), ping_request.clone()));
-
-        // Handle the ping and expect the disconnected Node to become connected
-        discv5.handle_rpc_response(enr2.node_id().clone(), 2, ping_response);
-        buckets = discv5.kbuckets.clone();
-
-        node = buckets.iter().next().unwrap();
-        assert_eq!(node.status, NodeStatus::Connected);
-    }
-
-    // The kbuckets table can have maximum 10 nodes in the same /24 subnet across all buckets
-    #[test]
-    fn test_table_limits() {
-        let keypair = identity::Keypair::generate_secp256k1();
-        let ip: IpAddr = "127.0.0.1".parse().unwrap();
-        let enr = EnrBuilder::new("v4")
-            .ip(ip.clone().into())
-            .udp(9000)
-            .build(&keypair)
-            .unwrap();
-        let mut discv5: Discv5<Box<u64>> =
-            Discv5::new(enr, keypair.clone(), ip.into(), true).unwrap();
-        let table_limit: usize = 10;
-        // Generate `table_limit + 1` nodes in the same subnet.
-        let enrs: Vec<Enr> = (1..=table_limit + 1)
-            .map(|i| {
-                let kp = identity::Keypair::generate_secp256k1();
-                let ip: IpAddr = IpAddr::V4(Ipv4Addr::new(192, 168, 1, i as u8));
-                EnrBuilder::new("v4")
-                    .ip(ip.clone().into())
-                    .udp(9000 + i as u16)
-                    .build(&kp)
-                    .unwrap()
-            })
-            .collect();
-        for enr in enrs {
-            discv5.add_enr(enr.clone());
-        }
-        // Number of entries could be < `table_limit` as per bucket limit is 2.
-        assert!(discv5.kbuckets_entries().collect::<Vec<_>>().len() <= table_limit);
-    }
-
-    // Each bucket can have maximum 2 nodes in the same /24 subnet
-    #[test]
-    fn test_bucket_limits() {
-        let keypair = identity::Keypair::generate_secp256k1();
-        let ip: IpAddr = "127.0.0.1".parse().unwrap();
-        let enr = EnrBuilder::new("v4")
-            .ip(ip.clone().into())
-            .udp(9500)
-            .build(&keypair)
-            .unwrap();
-        let bucket_limit: usize = 2;
-        // Generate `bucket_limit + 1` keypairs that go in `enr` node's 256th bucket.
-        let keypairs = {
-            let mut keypairs = Vec::new();
-            for _ in 0..bucket_limit + 1 {
-                loop {
-                    let keypair = identity::Keypair::generate_secp256k1();
-                    let enr_new = EnrBuilder::new("v4").build(&keypair).unwrap();
-                    let key: kbucket::Key<NodeId> = enr.node_id().clone().into();
-                    let distance = key
-                        .log2_distance(&enr_new.node_id().clone().into())
-                        .unwrap();
-                    if distance == 256 {
-                        keypairs.push(keypair);
-                        break;
-                    }
-                }
-            }
-            keypairs
-        };
-        // Generate `bucket_limit + 1` nodes in the same subnet.
-        let enrs: Vec<Enr> = (1..=bucket_limit + 1)
-            .map(|i| {
-                let kp = &keypairs[i - 1];
-                let ip: IpAddr = IpAddr::V4(Ipv4Addr::new(192, 168, 1, i as u8));
-                EnrBuilder::new("v4")
-                    .ip(ip.clone().into())
-                    .udp(9500 + i as u16)
-                    .build(kp)
-                    .unwrap()
-            })
-            .collect();
-        let mut discv5: Discv5<Box<u64>> =
-            Discv5::new(enr, keypair.clone(), ip.into(), true).unwrap();
-        for enr in enrs {
-            discv5.add_enr(enr.clone());
-        }
-
-        // Number of entries should be equal to `bucket_limit`.
-        assert_eq!(
-            discv5.kbuckets_entries().collect::<Vec<_>>().len(),
-            bucket_limit
-        );
-    }
-}
